Schedule UI improvement: Support popovers better
Not sure if this is an old bug or new browser behavior, but clicking on
a schedule item was showing the popover version but still jumping right
to its detail page. This should prevent that, still allowing the detail
link at the bottom of the popover to link to that page.

That said, some of the code around that is fairly inscrutable, so
there's definitely some room for improvement there anyway.