Rearrange some schedule UI elements
Instead of making the grid/list view toggle appear as tabs to the right
of the page, where they're easily overlooked and rather unclear, show
the links as options under the "View" tab. This also clarifies some of
the other options some. Hopefully.

Also, that stray href was apparently an edit left behind long ago, and
explains the behavior change worked around in the last commit. Seems
worth having those as href's, if only for search engine discovery
reasons, so leaving it as is for now.