Now compatible with Django 3.2 == LTS
This gets us to Django 3.2, but -should- also resolve any deprecation
issues that prevent going through 4.1. There's always those few things
that don't show up as DeprecationWarnings, but we can cross that bridge
when we get there.

(This is an backport of what was done to the MCFC prod code for that.)